Change Damage property from type int to short in Item class

+++ code/SpaceGameLibrary/SpaceGameLibrary/Item.cs
@@ -8,7 +8,7 @@ namespace SpaceGameLibrary
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Damage { get; set; }
+        public short Damage { get; set; }
         public short Value { get; set; }
         public bool Reusable { get; set; }
         public byte Quantity { get; set; }
